Hopefully fix publish-drafts scheduled workflow
According to this GitHub Discussion:
https://github.com/orgs/community/discussions/37103#discussioncomment-3964382
the actions/checkout action will configure the local git setup to use
its token. So this commit now tells the checkout action to use the token
that I set up in this repository's secrets, so that it persists into my
script so that the script triggers the GitHub Pages deployment.