Allow the use of '/' in comment fields
Previously, typing the slash key would jump to the search field, even if
the cursor was in one of the comment fields. The slash key will now no
longer jump to the search field if the focused element has the
`search-slash-ignore` attribute, which has also been applied to the
comment fields and the search field itself.