Show comment times in site timezone
This commit does a couple of things. It shows the time each comment was
made, in addition to the date. It also checks the localStorage to see
whether or not it should convert the comment into the local timezone.
Since comments currently have their dates stored in UTC, we make sure to
convert the comment date and time into either the site timezone or the
local timezone.

Closes #34